+++ src/datasources/data-frame/constants.ts
@@ -2,6 +2,11 @@ import { nbsp } from "./utils";
 
 export const decimationMethods = [
   {
+    value: 'NONE',
+    label: 'None',
+    description: nbsp`Data in full resolution without decimation`,
+  },
+  {
     value: 'LOSSY',
     label: 'Lossy',
     description: nbsp`Completes faster but is less accurate`,
@@ -17,12 +22,6 @@ export const decimationMethods = [
     description: nbsp`Maintains edges of data (includes max/min)`,
   },
 ];
-
-export const decimationNoneOption = {
-  value: 'NONE',
-  label: 'None',
-  description: nbsp`Data in full resolution without decimation`,
-};
 
 export const errorCodes: { [key: number]: string; } = {
   [-255134]: 'Invalid table ID',