+++ src/datasources/data-frame/types.ts
@@ -20,6 +20,11 @@ export type DataFrameDataQuery = DataFrameQueryV1 | DataFrameQueryV2;
 export type DataFrameVariableQuery = DataFrameQueryV1 | DataFrameVariableQueryV2;
 export type DataFrameQuery = DataFrameDataQuery | DataFrameVariableQuery;
 
+export interface PropertiesQueryCache {
+  requestInputs: string;
+  selectedProperties: string[];
+  response: TableProperties[];
+}
 
 export interface DataFrameQueryV1 extends DataQuery {
   type: DataFrameQueryType;
@@ -62,17 +67,7 @@ export const defaultVariableQueryV2: Omit<ValidDataFrameVariableQuery, 'refId'> 
   columnFilter: ''
 };
 
-export const defaultQueryV1: Omit<ValidDataFrameQueryV1, 'refId'> = {
-  type: DataFrameQueryType.Data,
-  tableId: '',
-  columns: [],
-  decimationMethod: 'LOSSY',
-  filterNulls: false,
-  applyTimeFilters: false
-};
-
 export const DataFrameFeatureTogglesDefaults: DataFrameFeatureToggles = {
-  queryUndecimatedData: false,
   highResolutionZoom: false
 };
 
@@ -261,11 +256,9 @@ export const metadataFieldOptions: Option[] = [
   { label: DATA_TABLE_NAME_LABEL, value: DATA_TABLE_NAME_FIELD, group: METADATA_GROUP }
 ];
 
-export type ValidDataFrameQuery = ValidDataFrameQueryV1 | ValidDataFrameQueryV2;
+export type ValidDataFrameQuery = ValidDataFrameQueryV2;
 
 export type ValidDataFrameVariableQuery = DataFrameVariableQueryV2 & Required<Omit<DataFrameVariableQueryV2, keyof DataQuery>>;
-
-export type ValidDataFrameQueryV1 = DataFrameQueryV1 & Required<Omit<DataFrameQueryV1, keyof DataQuery>>;
 
 export type ValidDataFrameQueryV2 = DataFrameQueryV2 & Required<Omit<DataFrameQueryV2, keyof DataQuery>>;
 
@@ -349,7 +342,6 @@ export interface TableDataRows {
 }
 
 export interface DataFrameFeatureToggles {
-  queryUndecimatedData: boolean;
   highResolutionZoom: boolean;
 }